+++ apps/grader-host/src/index.js
@@ -6,6 +6,7 @@ const tmp = require('tmp');
 const Docker = require('dockerode');
 const { S3 } = require('@aws-sdk/client-s3');
 const { SQSClient, SendMessageCommand } = require('@aws-sdk/client-sqs');
+const { ECRClient } = require('@aws-sdk/client-ecr');
 const { Upload } = require('@aws-sdk/lib-storage');
 const { exec } = require('child_process');
 const path = require('path');
@@ -14,10 +15,10 @@ const { pipeline } = require('node:stream/promises');
 const Sentry = require('@prairielearn/sentry');
 const sqldb = require('@prairielearn/postgres');
 const { sanitizeObject } = require('@prairielearn/sanitize');
-const { DockerName, setupDockerAuthAsync } = require('@prairielearn/docker-utils');
+const { DockerName, setupDockerAuth } = require('@prairielearn/docker-utils');
 
 const globalLogger = require('./lib/logger');
-const jobLogger = require('./lib/jobLogger');
+const { makeJobLogger } = require('./lib/jobLogger');
 const { config, loadConfig } = require('./lib/config');
 const healthCheck = require('./lib/healthCheck');
 const lifecycle = require('./lib/lifecycle');
@@ -25,6 +26,7 @@ const pullImages = require('./lib/pullImages');
 const receiveFromQueue = require('./lib/receiveFromQueue');
 const timeReporter = require('./lib/timeReporter');
 const load = require('./lib/load');
+const { makeAwsClientConfig, makeS3ClientConfig } = require('./lib/aws');
 
 const sql = sqldb.loadSqlEquiv(__filename);
 
@@ -106,7 +108,7 @@ async.series(
     },
     () => {
       globalLogger.info('Initialization complete; beginning to process jobs');
-      const sqs = new SQSClient({ region: config.awsRegion });
+      const sqs = new SQSClient(makeAwsClientConfig());
       for (let i = 0; i < config.maxConcurrentJobs; i++) {
         async.forever((next) => {
           if (!healthCheck.isHealthy() || processTerminating) return;
@@ -175,17 +177,12 @@ function isJobCanceled(job, callback) {
 function handleJob(job, done) {
   load.startJob();
 
-  const loggerOptions = {
-    bucket: job.s3Bucket,
-    rootKey: job.s3RootKey,
-  };
-
-  const logger = jobLogger(loggerOptions);
+  const logger = makeJobLogger();
   globalLogger.info(`Logging job ${job.jobId} to S3: ${job.s3Bucket}/${job.s3RootKey}`);
 
   const info = {
     docker: new Docker(),
-    s3: new S3({ region: config.awsRegion }),
+    s3: new S3(makeS3ClientConfig({ maxAttempts: 3 })),
     logger,
     job,
   };
@@ -249,7 +246,7 @@ async function reportReceived(info) {
       receivedTime: receivedTime,
     },
   };
-  const sqs = new SQSClient({ region: config.awsRegion });
+  const sqs = new SQSClient(makeAwsClientConfig());
   try {
     await sqs.send(
       new SendMessageCommand({
@@ -286,7 +283,8 @@ function initDocker(info, callback) {
       async () => {
         if (config.cacheImageRegistry) {
           logger.info('Authenticating to docker');
-          dockerAuth = await setupDockerAuthAsync(config.awsRegion);
+          const ecr = new ECRClient(makeAwsClientConfig());
+          dockerAuth = await setupDockerAuth(ecr);
         }
       },
       async () => {
@@ -453,14 +451,22 @@ function runJob(info, callback) {
             NetworkDisabled: !jobEnableNetworking,
             HostConfig: {
               Binds: [`${tempDir}:/grade`],
-              Memory: 1 << 30, // 1 GiB
-              MemorySwap: 1 << 30, // same as Memory, so no access to swap
-              KernelMemory: 1 << 29, // 512 MiB
-              DiskQuota: 1 << 30, // 1 GiB
+              Memory: config.graderDockerMemory,
+              MemorySwap: config.graderDockerMemorySwap,
+              KernelMemory: config.graderDockerKernelMemory,
+              DiskQuota: config.graderDockerDiskQuota,
               IpcMode: 'private',
-              CpuPeriod: 100000, // microseconds
-              CpuQuota: 90000, // portion of the CpuPeriod for this container
-              PidsLimit: 1024,
+              CpuPeriod: config.graderDockerCpuPeriod,
+              CpuQuota: config.graderDockerCpuQuota,
+              PidsLimit: config.graderDockerPidsLimit,
+              Ulimits: [
+                {
+                  // Disable core dumps, which can get very large and bloat our storage.
+                  Name: 'core',
+                  Soft: 0,
+                  Hard: 0,
+                },
+              ],
             },
             Entrypoint: entrypoint.split(' '),
           },
@@ -659,7 +665,7 @@ function uploadResults(info, callback) {
           messageBody.data = results;
         }
 
-        const sqs = new SQSClient({ region: config.awsRegion });
+        const sqs = new SQSClient(makeAwsClientConfig());
         await sqs.send(
           new SendMessageCommand({
             QueueUrl: config.resultsQueueUrl,
@@ -713,6 +719,17 @@ function uploadArchive(results, callback) {
             Bucket: s3Bucket,
             Key: `${s3RootKey}/archive.tar.gz`,
             Body: fs.createReadStream(tempArchive),
+          },
+        }).done();
+      },
+      async () => {
+        // Upload all logs to S3.
+        await new Upload({
+          client: s3,
+          params: {
+            Bucket: s3Bucket,
+            Key: `${s3RootKey}/output.log`,
+            Body: logger.getBuffer(),
           },
         }).done();
       },