+++ apps/prairielearn/python/zygote_utils_test.py
@@ -1,7 +1,34 @@
+import json
 from typing import Any
 
 import pytest
 import zygote_utils as zu
+
+
+@pytest.mark.parametrize(
+    "item", ["-9007199254740991", "-1", "0", "1", "9007199254740991"]
+)
+def test_safe_parse_int_small_ints(item: str) -> None:
+    loaded_item = json.loads(item, parse_int=zu.safe_parse_int)
+    assert isinstance(loaded_item, int)
+    assert int(item) == loaded_item
+
+
+@pytest.mark.parametrize(
+    "item",
+    [
+        "-2.8e16",
+        "28000000000000000",
+        "-9007199254740992",
+        "9007199254740992",
+        "28000000000000000",
+        "2.8e16",
+    ],
+)
+def test_safe_parse_int_large_ints(item: str) -> None:
+    loaded_item = json.loads(item, parse_int=zu.safe_parse_int)
+    assert isinstance(loaded_item, float)
+    assert float(item) == loaded_item
 
 
 @pytest.mark.parametrize(