+++ packages/flash/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @prairielearn/flash
 
+## 1.1.1
+
+### Patch Changes
+
+- 8dd894623: Upgrade all dependencies
+- Updated dependencies [8dd894623]
+  - @prairielearn/html@3.0.9
+
+## 1.1.0
+
+### Minor Changes
+
+- 0f5ef9c18: Add support for html-safe flash messages
+
+## 1.0.6
+
+### Patch Changes
+
+- bd0053577: Upgrade all dependencies
+
+## 1.0.5
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+
 ## 1.0.4
 
 ### Patch Changes