+++ packages/postgres-tools/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @prairielearn/postgres-tools
 
+## 1.1.9
+
+### Patch Changes
+
+- 8dd894623: Upgrade all dependencies
+- Updated dependencies [8dd894623]
+  - @prairielearn/postgres@1.7.8
+
+## 1.1.8
+
+### Patch Changes
+
+- bd0053577: Upgrade all dependencies
+- Updated dependencies [bd0053577]
+  - @prairielearn/postgres@1.7.7
+
+## 1.1.7
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+- Updated dependencies [f03853d90]
+  - @prairielearn/postgres@1.7.6
+
 ## 1.1.6
 
 ### Patch Changes